Added "revision toolbar" - special "button" insertable into any place in toolbar,  allow to list and change revision, added ability to put labels in toolbars Revision switching does not fully work yet in GUI  (something crashes, something does not work)
Added keyboard shortcut for changing pages

Fixed some issued with scripting classes.

Fixed localization issues (not loading loc. files, not localizing toolbar texts)

+++ src/gui/qsarray.h
@@ -33,7 +33,7 @@ public slots:
  /*- Append string element at end of array */
  void add(const QString &ip);
  /*- Return size of this array (number of elements) */
- size_t count();
+ int count();
  /*- Delete element with given index from this array. Elements with higher index (if any) are shifted to occupy the space */
  void delProperty(int index);
  /*- Return string representation of this array */
